requirements.txt
absl-py==2.1.0
accelerate==0.30.1
adjustText==1.2.0
aiohttp==3.9.5
aiosignal==1.3.1
airr==1.5.1
anarci==1.3
anndata==0.10.8
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
array_api_compat==1.7.1
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
awkward==2.6.6
awkward-cpp==35
Babel==2.15.0
beautifulsoup4==4.12.3
biopython==1.83
bleach==6.1.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
Cython==0.29.37
datasets==2.19.1
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
exceptiongroup==1.2.1
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.14.0
flatbuffers==24.3.25
fonttools==4.52.4
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.3.1
future==1.0.0
gast==0.5.4
google-auth==2.30.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphtools==1.5.3
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
hdbscan==0.8.36
hmmlearn==0.3.2
httpcore==1.0.5
httpx==0.27.0
huggingface==0.0.1
huggingface-hub==0.23.2
idna==3.7
igraph==0.11.5
importlib_metadata==8.0.0
ipykernel==6.29.5
ipython==8.24.0
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
kaleido==0.2.1
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
legacy-api-wrap==1.4
leidenalg==0.10.2
Levenshtein==0.25.1
libclang==18.1.1
lightning-utilities==0.11.2
llvmlite==0.42.0
logomaker==0.8
magic-impute==3.0.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mhcgnomes==1.8.6
mhcnames==0.4.8
mistune==3.0.2
ml-dtypes==0.3.2
mpmath==1.3.0
mudata==0.2.4
multidict==6.0.5
multiprocess==0.70.16
muon==0.1.6
namex==0.0.8
natsort==8.4.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
notebook==7.2.1
notebook_shim==0.2.4
numba==0.59.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opt-einsum==3.3.0
optree==0.11.0
overrides==7.7.0
packaging==24.0
pandas==1.5.3
pandocfilters==1.5.1
parasail==1.3.4
parso==0.8.4
patsy==0.5.6
pdf2image==1.17.0
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
plotly==5.24.1
pooch==1.8.2
prometheus_client==0.20.0
prompt_toolkit==3.0.44
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
Pygments==2.18.0
PyGSP==0.5.1
pyhmmer==0.10.12
pynndescent==0.5.12
pyparsing==3.1.2
pyreadr==0.5.0
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-Levenshtein==0.25.1
python-version==0.0.2
pytorch-lightning==2.2.5
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
rapidfuzz==3.9.2
referencing==0.35.1
regex==2024.5.15
requests==2.32.2
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.19.0
rpy2==3.5.16
rsa==4.9
safetensors==0.4.3
scanpy==1.10.3
scikit-learn==1.5.0
scipy==1.13.1
scirpy==0.17.2
scprep==1.2.3
seaborn==0.13.2
Send2Trash==1.8.3
serializable==0.4.1
session-info==1.0.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
squarify==0.4.3
stack-data==0.6.3
statannotations==0.6.0
statsmodels==0.15.0.dev514+gcb252755e
stdlib-list==0.10.0
sympy==1.12
tasklogger==1.2.0
tenacity==9.0.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
terminado==0.18.1
texttable==1.7.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.5.0
tinycss2==1.3.0
tokenizers==0.19.1
tomli==2.0.1
torch==2.3.0
torchmetrics==1.4.0.post0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers @ git+https://github.com/huggingface/transformers@9d35edbb30625489bf286a9b15aed0c5a3119c1c
triton==2.3.0
typechecks==0.1.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.0
tzdata==2024.1
tzlocal==5.2
umap==0.1.1
umap-learn==0.5.6
uri-template==1.3.0
urllib3==2.2.1
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
wrapt==1.16.0
xxhash==3.4.1
yamlordereddictloader==0.4.2
yarl==1.9.4
zipp==3.19.2
